rate-track: write track metadata to file
In addition to adding the track rating to the mpd stickers database,
rate-track now writes the track rating to the FMPS_RATING tag of flac
and mp3 files as per the Free Media Player Specification.

It uses `metaflac` and `mid3v2` to do these if they are found in the
user's `$PATH`. If not, it skips them.